requirements.txt
dnspython>=2.6.1,<3.0.0
nagiosplugin>=1.3.3,<2.0.0
